Create PrimitiveTypesRecap-31
Premitive type
   // byte   // short     // int      // long        // float      // double        // char       // boolean

non-Premitive
String/array/etc.
Java String. Basically, string is a sequence of characters but it's not a primitive type. When we create a string in java, it actually creates an object of type String. String is immutable object which means that it cannot be changed once it is created.