Rename Operators,Operand&Expressionsinmoredetail-32.java to Operators,Operand&Expressionsinmoredetail32.java
Java defines several bitwise operators, which can be applied to the integer types, long, int, short, char, and byte. Binary AND Operator copies a bit to the result if it exists in both operands. ... Binary Left Shift Operator. The left operands value is moved left by the number of bits specified by the right operand.